+++ hw2/hw2/InsertionSort.cpp
@@ -13,7 +13,7 @@ CInsertionSort::~CInsertionSort()
 
 void CInsertionSort::sort(std::vector<int>& vector)
 {
-	for (unsigned int i = 1; i < vector.size(); i++) {
+	for (size_t i = 1; i < vector.size(); i++) {
 		int j = i;
 		while ((j > 0) && (vector[j - 1] > vector[j])) {
 			int tmp = vector[j - 1];
