+++ hw2/hw2/SortInterface.h
@@ -7,5 +7,5 @@ struct ISortInterface {
 	virtual ~ISortInterface() {}
 	virtual void sort(std::vector<int>& vector) = 0; 
 	virtual std::string identify() { return ""; };
-	void display(std::vector<int>& vector) { for (unsigned int i = 0; i < vector.size(); i++) { std::cout << i << " " << vector[i] << std::endl; } };
+	void display(std::vector<int>& vector) { for (size_t i = 0; i < vector.size(); i++) { std::cout << i << " " << vector[i] << std::endl; } };
 };
